--- portfolio/components/project-menu.tsx
@@ -1,11 +1,0 @@
-const ProjectMenu = () => {
-  return (
-    <section className="flex-col md:flex-row flex items-center md:justify-between mt-16 mb-16 md:mb-12">
-      <h2 className="md:text-2xl font-bold tracking-tighter leading-tight md:pr-8">
-        view selected work, or filter by ...
-      </h2>
-    </section>
-  )
-}
-
-export default ProjectMenu
